+++ src/bika/lims/browser/analysisrequest/view.py
@@ -22,6 +22,9 @@ from bika.lims import api
 from bika.lims.browser import BrowserView
 from bika.lims.browser.header_table import HeaderTableView
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
+from senaite.core.interfaces import ISampleSection
+from zope.component import subscribers
+
 from resultsinterpretation import ARResultsInterpretationView
 
 
@@ -47,39 +50,13 @@ class AnalysisRequestViewView(BrowserView):
 
         return self.template()
 
-    def render_analyses_table(self, table="lab"):
-        """Render Analyses Table
+    def get_sections(self):
+        """Returns a list with adapters that implement ISampleSection
         """
-        if table not in ["lab", "field", "qc"]:
-            raise KeyError("Table '{}' does not exist".format(table))
-        view_name = "table_{}_analyses".format(table)
-        view = api.get_view(
-            view_name, context=self.context, request=self.request)
-        # Call listing hooks
-        view.update()
-        view.before_render()
-        return view.ajax_contents_table()
-
-    def has_lab_analyses(self):
-        """Check if the AR contains lab analyses
-        """
-        # Negative performance impact - add a Metadata column
-        analyses = self.context.getAnalyses(getPointOfCapture="lab")
-        return len(analyses) > 0
-
-    def has_field_analyses(self):
-        """Check if the AR contains field analyses
-        """
-        # Negative performance impact - add a Metadata column
-        analyses = self.context.getAnalyses(getPointOfCapture="field")
-        return len(analyses) > 0
-
-    def has_qc_analyses(self):
-        """Check if the AR contains field analyses
-        """
-        # Negative performance impact - add a Metadata column
-        analyses = self.context.getQCAnalyses()
-        return len(analyses) > 0
+        # We use subscriber adapters here because we need different add-ons to
+        # be able to add their own sections without dependencies amongst them
+        sections = subscribers((self.context, ), ISampleSection)
+        return sorted(sections, key=lambda s: getattr(s, "order", 100))
 
     def is_hazardous(self):
         """Checks if the AR is hazardous